backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
openai
httpx
google-genai